requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.12.14
aioresponses==0.7.7
aiosignal==1.3.2
art==6.4
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
dnspython==2.7.0
frozenlist==1.5.0
httpretty==1.1.4
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
Pygments==2.18.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-Wappalyzer==0.3.1
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
setuptools==78.1.1
soupsieve==2.6
urllib3==2.5.0
yarl==1.18.3